Fix for #34962 - On CSV import, product variations are not visible...
Fix incorrect behaviour (return of float for integer value) from 7e3eec3ed6e488c63637b409b4ac846a3ff813e4, which causes problems with code introduced in 37777432d42a15fb2a2e725f980feb5aa92a6e9b.

Changes kept to a minimum for compatibility.

+++ plugins/woocommerce/includes/import/class-wc-product-csv-importer.php
@@ -657,7 +657,7 @@ class WC_Product_CSV_Importer extends WC_Product_Importer {
 	 *
 	 * @param string $value Field value.
 	 *
-	 * @return float|string
+	 * @return int|string
 	 */
 	public function parse_published_field( $value ) {
 		if ( '' === $value ) {
@@ -671,7 +671,7 @@ class WC_Product_CSV_Importer extends WC_Product_Importer {
 			return wc_string_to_bool( $value ) ? 1 : -1;
 		}
 
-		return floatval( $value );
+		return intval( $value );
 	}
 
 	/**
